+++ lib/component/gradation_box.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class GradationBox extends StatelessWidget {
+  const GradationBox({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: 35,
+      width: 250,
+      child: Container(
+        decoration: const BoxDecoration(
+          gradient: LinearGradient(
+            colors: [Colors.black, Color.fromARGB(127, 0, 195, 255)],
+            begin: Alignment.centerLeft,
+            end: Alignment.centerRight,
+          ),
+        ),
+      ),
+    );
+  }
+}
