+++ lib/component/space_box.dart
