--- lib/view/select_bank/pad_button_area_view.dart
@@ -1,44 +1,0 @@
-// Flutter imports:
-import 'package:flutter/material.dart';
-
-// Project imports:
-import 'package:simple_sample_manager/constant/page_back_button.dart';
-
-// Project imports:
-
-class PadButtonArea extends StatelessWidget {
-  const PadButtonArea({Key? key}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        const SizedBox(
-          height: 270,
-          width: 350,
-        ),
-        Padding(
-          padding: const EdgeInsets.only(right: 40, left: 40),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              const PageBackButton(),
-              SizedBox(
-                height: 35,
-                width: 250,
-                child: Container(
-                  decoration: const BoxDecoration(
-                    gradient: LinearGradient(
-                      colors: [Colors.black, Colors.blue],
-                      begin: Alignment.centerLeft,
-                      end: Alignment.centerRight,
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-}
