+++ lib/view_model/top_bank_button_view_model.dart
@@ -2,7 +2,7 @@
 import 'package:flutter/material.dart';
 
 // Project imports:
-import 'package:simple_sample_manager/constant/bank_button.dart';
+import 'package:simple_sample_manager/component/bank_button.dart';
 
 class BankButtonArea extends StatelessWidget {
   const BankButtonArea({Key? key}) : super(key: key);
